IGAPP-970: Merge CitySelector and FilterableCitySelector and extract NearbyCities
