+++ native/src/components/CitySelector.tsx
@@ -1,49 +1,38 @@
 import { groupBy, transform } from 'lodash'
-import React, { ReactElement, ReactNode } from 'react'
+import React, { ReactElement, ReactNode, useState } from 'react'
 import { TFunction } from 'react-i18next'
 import { View } from 'react-native'
-import { Button } from 'react-native-elements'
-import Icon from 'react-native-vector-icons/MaterialIcons'
 import styled, { useTheme } from 'styled-components/native'
 
 import { CityModel, filterSortCities } from 'api-client'
 
 import buildConfig from '../constants/buildConfig'
-import { LocationInformationType } from '../hooks/useUserLocation'
-import getNearbyCities from '../utils/getNearbyCities'
 import CityEntry from './CityEntry'
 import CityGroup from './CityGroup'
+import NearbyCities from './NearbyCities'
 import NothingFound from './NothingFound'
+import SearchInput from './SearchInput'
 
 const CityGroupContainer = styled.View`
   flex: 0;
   flex-direction: column;
 `
-const NearbyMessageContainer = styled.View`
-  padding: 7px;
+
+const SearchBar = styled.View`
   flex-direction: row;
-  justify-content: space-between;
-`
-const RetryButtonContainer = styled.View`
-  flex-direction: column;
-  height: 46px;
-`
-const NearbyMessage = styled.Text`
-  color: ${props => props.theme.colors.textColor};
-  font-family: ${props => props.theme.fonts.native.decorativeFontRegular};
-  padding-top: 15px;
+  align-items: center;
+  justify-content: center;
+  padding: 0 10%;
 `
 
 type PropsType = {
   cities: Array<CityModel>
-  filterText: string
   navigateToDashboard: (city: CityModel) => void
-  locationInformation: LocationInformationType
   t: TFunction<'landing'>
 }
 
-const CitySelector = ({ cities, filterText, navigateToDashboard, locationInformation, t }: PropsType): ReactElement => {
-  const { status, coordinates, message, requestAndDetermineLocation } = locationInformation
+const CitySelector = ({ cities, navigateToDashboard, t }: PropsType): ReactElement => {
+  const [filterText, setFilterText] = useState<string>('')
   const theme = useTheme()
 
   const resultCities = filterSortCities(cities, filterText, buildConfig().featureFlags.developerFriendly)
@@ -76,43 +65,23 @@ const CitySelector = ({ cities, filterText, navigateToDashboard, locationInforma
     []
   )
 
-  const nearbyCitiesErrorMessage = (
-    <NearbyMessage>{t(status === 'unavailable' ? message : 'noNearbyCities')}</NearbyMessage>
-  )
-  const nearbyCities =
-    coordinates &&
-    getNearbyCities(
-      cities.filter(city => city.live),
-      coordinates[0],
-      coordinates[1]
-    )
-  const retryButton = status === 'unavailable' && (
-    <RetryButtonContainer>
-      <Button
-        icon={<Icon name='refresh' size={30} color={theme.colors.textSecondaryColor} />}
-        title=''
-        type='clear'
-        onPress={requestAndDetermineLocation}
-        accessibilityLabel={t('refresh')}
-        accessibilityRole='button'
-      />
-    </RetryButtonContainer>
-  )
-
   return (
     <View>
-      <CityGroupContainer>
-        <CityGroup>{t('nearbyCities')}</CityGroup>
-        {nearbyCities?.length ? (
-          nearbyCities.map(renderCity)
-        ) : (
-          <NearbyMessageContainer>
-            {nearbyCitiesErrorMessage}
-            {retryButton}
-          </NearbyMessageContainer>
-        )}
-      </CityGroupContainer>
-      {cityEntries}
+      <SearchBar>
+        <SearchInput
+          filterText={filterText}
+          onFilterTextChange={setFilterText}
+          placeholderText={t('searchCity')}
+          spaceSearch={false}
+        />
+      </SearchBar>
+      <View>
+        <CityGroupContainer>
+          <CityGroup>{t('nearbyCities')}</CityGroup>
+          <NearbyCities cities={cities} navigateToDashboard={navigateToDashboard} filterText={filterText} t={t} />
+        </CityGroupContainer>
+        {cityEntries}
+      </View>
     </View>
   )
 }
